refactor: More shifting around
