+++ arcade-events-server/src/main/java/net/casual/arcade/events/server/mixins/ScoreboardMixin.java
@@ -14,8 +14,8 @@ import net.casual.arcade.events.server.entity.EntityTeamLeaveEvent;
 import net.casual.arcade.events.server.player.PlayerTeamJoinEvent;
 import net.casual.arcade.events.server.player.PlayerTeamLeaveEvent;
 import net.casual.arcade.utils.PlayerUtils;
-import net.casual.arcade.utils.StringUtilsKt;
 import net.casual.arcade.utils.entity.EntityUtilsKt;
+import net.casual.arcade.utils.string.StringUtilsKt;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.server.ServerScoreboard;
 import net.minecraft.server.level.ServerPlayer;
