+++ arcade-utils/src/main/kotlin/net/casual/arcade/utils/string/StringUtils.kt
@@ -2,11 +2,10 @@
  * Copyright (c) 2024 senseiwells
  * Licensed under the MIT License. See LICENSE file in the project root for details.
  */
-package net.casual.arcade.utils
+package net.casual.arcade.utils.string
 
-import net.casual.arcade.utils.string.StringCasingDecoder
-import net.casual.arcade.utils.string.StringCasingEncoder
 import java.util.*
+import kotlin.text.iterator
 
 private val SMALL_CAPS_ALPHABET = "ᴀʙᴄᴅᴇꜰɢʜɪᴊᴋʟᴍɴᴏᴘqʀꜱᴛᴜᴠᴡxyᴢ".toCharArray()
