+++ arcade-visuals/src/main/kotlin/net/casual/arcade/visuals/utils/elements/component/TPSComponentElement.kt
@@ -6,7 +6,7 @@ package net.casual.arcade.visuals.utils.elements.component
 
 import net.casual.arcade.utils.ColorUtils
 import net.casual.arcade.utils.component.color
-import net.casual.arcade.utils.calculateTPS
+import net.casual.arcade.utils.server.calculateTPS
 import net.casual.arcade.visuals.elements.UniversalElement
 import net.minecraft.network.chat.Component
 import net.minecraft.server.MinecraftServer
